refactor: move geolocation banner in components folder
